상품등록 성공

+++ src/main/java/com/travel/travel/Repository/ItemRepository.java
@@ -5,11 +5,10 @@ import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import org.springframework.data.querydsl.binding.QuerydslPredicate;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> ,
+public interface ItemRepository extends JpaRepository<Item, Long> ,
         QuerydslPredicateExecutor<Item> {
     List<Item> findByItemNm(String name);
